feat(x509): Add get_tbs function to common module
Move the get_tbs function from multiple locations into
caliptra_common::x509 module to avoid code duplication. The function
extracts the TBS (To Be Signed) portion from a DER encoded X509
certificate.
