+++ rom/dev/Cargo.toml
@@ -30,6 +30,7 @@ zeroize.workspace = true
 memoffset.workspace = true
 
 [build-dependencies]
+caliptra_common = { workspace = true, default-features = false, features = ["std"] }
 cfg-if.workspace = true
 const-gen.workspace = true
 hex.workspace = true
