+++ runtime/Cargo.toml
@@ -46,6 +46,7 @@ aes.workspace = true
 aes-gcm.workspace = true
 caliptra-api.workspace = true
 caliptra-builder.workspace = true
+caliptra_common = { workspace = true, default-features = false, features = ["std"] }
 caliptra-emu-bus.workspace = true
 caliptra-hw-model.workspace = true
 caliptra-image-elf.workspace = true
